refactor(e2e): align Solana suite layout with Tron structure (Phase 1)
